server/requirements.txt
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
colorama==0.4.3
comm==0.2.2
contourpy==1.1.1
cryptography==2.8
cycler==0.12.1
debugpy==1.8.6
decorator==4.4.2
defusedxml==0.7.1
pymongo==4.5.0
python-dotenv==1.0.0